Implement general message to contract predicate (#11)
## Project Layout
This PR both changes the script and predicate to a more generic format
as well as adds tests. There are four main parts to this repo:

### contract-message-predicate
This is the main part of the whole process that verifies the conditions
specified in the README. Contains a _utils.sw_ file to help with GTF
related calls (eventually will be added to sway std-lib).

### contract-message-receiver
This is a simple interface abi that any contract that wants to receive
message must implement. Also contains a _utils.sw_ file to help with GTF
related calls.

### contract-message-script
This is a simple script that makes a single call to a contract to
process a message. Also contains a _utils.sw_ file to help with GTF
related calls.

### contract-message-test
This part contains both the actual tests as well as some sway code with
a message receiving example contract used in the tests. Also contains a
_utils.sw_ file to help with GTF related calls.

#### Tests
_**harness.rs**_ -> The testing entry point which checks three success
cases with different allowable options followed by several fail cases
that should result in an invalid predicate based on the conditions
specified in the README.
_**utils/environment.rs**_ -> The main helper util which contains quick
functions for starting up a test client with a given starting
messages/coin state to test with. It also contains some smaller helpful
functions intended to reduce redundant code in the test cases.
_**utils/builder.rs**_ -> This helper util adds support for building a
transaction that relays a message to a contract using the standard
script and predicate. Deals with anything related to the script and
predicate bytecode.

Co-authored-by: Braqzen <103777923+Braqzen@users.noreply.github.com>