Updated nomenclature (#15)
A few items in the codebase use the term "sidechain", since the current
design is geared towards an initial stepping stone that behaves more
like a sidechain than a rollup or L2. This should be replaced with the
just the terminology of "FuelChain" since the contracts are now
upgradeable and may become more than just this initial design.