Bump forc to 0.56, rust sdk to 0.60, adapt predicate to new encoding (#180)
This PR updates dependencies with the current target of testnet

NOTICE: due to the enforced requirement to use the new v1 encoding, the
predicate has been substantially modified, as the calldata that gets
passed to the sway bridge now follows a different pattern.

During this process and according to different audits that have
signalled that ETH values on the payload could get potentially stuck on
the sway bridge, I have decided to go beyond and completely remove any
fund forwarding of the predicate to the bridge. If it is going to get
stuck anyway, we may as well simplify its logic so that the funds do
stay at the predicate, not at the bridge. This issue has been cataloged
as informational / low risk because in order to happen, the solidity
contracts need to be flawed themselves.

---------

Co-authored-by: K1-R1 <77465250+K1-R1@users.noreply.github.com>